Update day 1 for SPASM and the embedded version
Most notably this replaces the `b_call` macro with `bcall` because that's
how the include files in use spell it. This fixes a lot of the partial editing
that had been done on that day, replacing attempts to be cross-platform with
assuming you'll use your browser and providing some cursory documentation for
the case of not using browser-based tools.